requirements.txt
django
gunicorn
whitenoise
python-dotenv
dj-database-url
psycopg2-binary